+++ frontend/src/components/Dashboard.js
@@ -1,54 +1,21 @@
 import React from "react";
 import { useState, useEffect, useContext } from "react";
-import GameCard from "./GameCard";
 import { Link } from "react-router-dom";
-import { getGames } from "../functions/Fetch";
 import { countGames } from "../utils/sanity/gameServices";
-import MyFavourites from "./MyFavourites";
+import MyFavourites from "./dashboardComponents/MyFavourites";
 import UserContext from "../context/UserContext";
-import { fetchGamesByUsername } from "../utils/sanity/userServices";
+import MyGames from "./dashboardComponents/MyGames";
+import GameShop from "./dashboardComponents/GameShop";
 
 
 export default function Dashboard() {
-  //state for spill fra Sanity
-  const [userGames, setUserGames] = useState([]);
+
   //state for antall spill i MyGames
   const [numGames, setNumGames] = useState(0);
 
-  const [games, setGames] = useState([]);
   //const count = 0;
 
   const { user } = useContext(UserContext);
-
-  useEffect(() => {
-    getGames({
-      ordering: "-released",
-      page_size: 3,
-    }).then((results) => {
-      setGames(results);
-    });
-  }, []);
-
-  //Henter spill fra Sanity....
-  async function getUserGames(username) {
-    const data = await fetchGamesByUsername(username);
-    const userGames = data.games
-    return userGames.slice(0,4)
-  } 
-
-   useEffect(() => {
-    if (user) {
-      getUserGames(user.username)
-        .then((userGames) => {
-         
-          setUserGames(userGames);
-        })
-        .catch((error) => {
-          console.error(error);
-          window.location.href = "/login";
-        })
-    }
-  }, [user]);
 
   async function getCount() {
     const data = await countGames();
@@ -73,21 +40,8 @@ export default function Dashboard() {
             Visit shop
           </Link>
         </section>
-        <article className="gs-box">
-          {games?.map((game, index) => (
-            <GameCard
-              key={index}
-              image={game.background_image}
-              slug={game.slug}
-              title={game.name}
-              playTime={game.playtime}
-              genre={game.genres.map((genre, index) => (
-                <li key={index}>{genre.name}</li>
-              ))}
-              text="Buy"
-            />
-          ))}
-        </article>
+        <GameShop />
+    
       </section>
       {user ? (
         <>
@@ -104,26 +58,8 @@ export default function Dashboard() {
                 My Games
               </Link>
             </section>
+            <MyGames />
 
-          {userGames !== undefined ? ( 
-            <article className="gs-box">
-          
-              {userGames?.map((item, index) => (
-                  <GameCard
-                    key={index}
-                    title={item.game.title}
-                    genre={item.game.genres.map((genre, index) => (
-                      <li key={index}>{genre.title}</li>
-                    ))}
-                    image={item.game.image}
-                    slug={item.slug}
-                    playTime={item.playtime}
-                
-                    text="Buy"
-                  />
-              ))}
-
-            </article>) : ('')}
           </section>
           <section className="gs">
             <section className="gs-txt-box">
@@ -137,6 +73,7 @@ export default function Dashboard() {
               </Link>
             </section>
             <MyFavourites />
+              
           </section>
         </>
       ) : (
